feat(chat-sidebar): refactor chat sidebar logic into a custom hook for better organization and reusability feat(preview-panel): refactor preview panel logic into a custom hook to streamline functionality and improve maintainability chore(types): add TypeScript types for chat sidebar and preview panel hooks to enhance type safety and developer experience style(justfile): add a new command to remove preview containers for better development workflow
